+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 2.0.0 (11 April 2019)
+
+- Add `canReachNotchAt` to disable or enable notches on the fly
+- OverlayContainer now uses the last view controller of its viewControllers as the overlay view controller. It stacks the other view controllers on top of each other, if any, and adds them underneath the overlay view controller.
+- Handle a new edge case: the overlay container has only one notch
+
+## 1.3.0 (04 April 2019)
+
+- Add completion block to `moveOverlay`
+
+## 1.2.0 (27 March 2019)
+
+- Hide `UIViewController+Children` methods
+
+## 1.1.2 (27 March 2019)
+
+- Fix Xcode 10.2 compilation (Nimble & Quick update)
+- Hide Carthage test dependancies
+
 ## 1.1.1 (27 February 2019)
 
 - Fix the cleaning of the translation drivers when a new driving scroll view is set.